fix(pg): cast auto_queue_entries.thread_group/batch_phase to BIGINT in slot thread name lookup
Discord delivery's build_slot_thread_name_pg decoded thread_group/batch_phase
as i64 but the PG columns are INT4. Each dispatch_outbox notify entry hit
"mismatched types: i64 vs INT4" and got stuck on retries (4/4 entries pending
after 4 minutes — no Discord messages delivered for the auto-queue dispatch).

Adds explicit ::BIGINT cast to two SELECT statements in
src/server/routes/dispatches/discord_delivery.rs:1576, 1600 — same pattern
as a79500c6 (metadata JSONB cast).

Other PG sites that already cast were untouched. Outstanding similar
regressions exist for tokens (INT4 vs i64) in dispatched_sessions and
column a.cli_provider missing — separately tracked.

+++ src/server/routes/dispatches/discord_delivery.rs
@@ -1573,7 +1573,7 @@ async fn build_slot_thread_name_pg(
 ) -> Result<String, String> {
     let mut batch_phase_for_label = 0i64;
     let group_info = sqlx::query(
-        "SELECT run_id, COALESCE(thread_group, 0) AS thread_group, COALESCE(batch_phase, 0) AS batch_phase
+        "SELECT run_id, COALESCE(thread_group, 0)::BIGINT AS thread_group, COALESCE(batch_phase, 0)::BIGINT AS batch_phase
          FROM auto_queue_entries
          WHERE dispatch_id = $1
          LIMIT 1",
@@ -1597,7 +1597,7 @@ async fn build_slot_thread_name_pg(
     .transpose()?
     .or(
         sqlx::query(
-            "SELECT run_id, COALESCE(thread_group, 0) AS thread_group, COALESCE(batch_phase, 0) AS batch_phase
+            "SELECT run_id, COALESCE(thread_group, 0)::BIGINT AS thread_group, COALESCE(batch_phase, 0)::BIGINT AS batch_phase
              FROM auto_queue_entries
              WHERE kanban_card_id = $1
                AND status IN ('pending', 'dispatched')
